[MF-196, MF-225, MF-229, MF-230, ] В статусе отправлено 1970 год (Статусы Черновик и Отправлено, с временем 1970), на сервере не записывается причина перевода статуса, при старте приложения - оно теперь обновляет данные 2 раза, один штатный, один из-за несоответствия версиив резолюциях - обвести границу и написать - резолюция. https://mobiforceteam.atlassian.net/browse/MF-196 https://mobiforceteam.atlassian.net/browse/MF-225 https://mobiforceteam.atlassian.net/browse/MF-229 https://mobiforceteam.atlassian.net/browse/MF-230

+++ lib/data/models/resolution_model.dart
@@ -23,7 +23,7 @@ class ResolutionModel extends ResolutionEntity
     return TaskModel(id: int.parse(json["id"]??0), name: json["name"]??"", address: json["address"]??"", client: json["client"]??"", subdivision: json["subdivision"]??"");
   }*/
 
-  ResolutionModel({required id,required usn,required serverId,required name, client, color, required resolutionGroup}): super(
+  ResolutionModel({required id,required usn,required serverId,required name, client, color, resolutionGroup}): super(
       id:id,
       usn:usn,
       serverId:serverId,
@@ -51,13 +51,15 @@ class ResolutionModel extends ResolutionEntity
       //print ("db id == ${t.toString()}");
     }
 
-    await db.deleteResolutuionGroupRelation(t.id);
-    Future.forEach(resolutionGroup, (ResolutionGroupModel element) async {
-      int resolutionGroupId = await element.insertToDB(db);
-      await db.insertResolutuionGroupRelation(t.id,resolutionGroupId);
-    });
+    if(resolutionGroup!=null) {
+      await db.deleteResolutuionGroupRelation(t.id);
+      Future.forEach(resolutionGroup!, (ResolutionGroupModel element) async {
+        int resolutionGroupId = await element.insertToDB(db);
+        await db.insertResolutuionGroupRelation(t.id, resolutionGroupId);
+      });
+    }
     //print ("db id == ${t.id}");
-    return 0;
+    return t.id;
   }
   factory ResolutionModel.fromMap({required Map<String, dynamic> map, List<Map<String, dynamic>> mapResolutionsList = const[]})
   {
@@ -82,7 +84,7 @@ class ResolutionModel extends ResolutionEntity
         serverId: int.parse(json["id"]??"0"),
         name: json["name"]??"",
         color: json["color"]??"",
-        resolutionGroup:json["resolutionGroups"]!=null?(json["resolutionGroups"] as List).map((person) => ResolutionGroupModel.fromJson(person)).toList():<ResolutionGroupModel>[],
+        resolutionGroup:json["resolutionGroups"]!=null?(json["resolutionGroups"] as List).map((person) => ResolutionGroupModel.fromJson(person)).toList():null,
         //client: json["client"]??"",
         //address: json["address"]??""
     );
